pulls in fixes from issue #38 and issue #39
final fixes from branch before merge to main
later plan to merge algoria-search directly into main

+++ scripts/clean_rebuild.sh
@@ -12,12 +12,6 @@ while getopts "d:" option; do
       d) # set build dir
         ARG_BUILD_DIR=${OPTARG}
         ;;
-      i) # set identity
-        ARG_ID_FILE=${OPTARG}
-        ;;
-      h) # set host
-        ARG_HOST+=("${OPTARG}")
-        ;;
       :) # no args
         Help;
         ;;
@@ -27,16 +21,16 @@ while getopts "d:" option; do
    esac
 done
 # check for required args
-if [ -z $ARG_BUILD_DIR ]; then
+if [ -z "$ARG_BUILD_DIR" ]; then
   echo -e "Missing required argument -d directory"
   Help;
 fi
 
 # create dir if it does not exist
-[ ! -d $ARG_BUILD_DIR ] && mkdir -p $ARG_BUILD_DIR
+[ ! -d "$ARG_BUILD_DIR" ] && mkdir -p "$ARG_BUILD_DIR"
 
 # remove everthing under the build dir
-rm -rf $ARG_BUILD_DIR/*
+rm -rf "${ARG_BUILD_DIR:?}"/* || exit
 # remove working directories
 rm -rf ../working/*
 
@@ -46,15 +40,17 @@ for gitrepo in eosnetworkfoundation/welcome \
     AntelopeIO/leap \
     AntelopeIO/DUNE \
     eosnetworkfoundation/mandel-eosjs \
-    eosnetworkfoundation/mandel-java \
-    eosnetworkfoundation/mandel-swift
+    eosnetworkfoundation/mandel-java 
 do
   echo "working on ${gitrepo}"
   # empty out var
   unset branch
-  if [ -z $branch ]; then
-    ./generate_documents.sh -d $ARG_BUILD_DIR -r ${gitrepo}
+  if [ -z "$branch" ]; then
+    ./generate_documents.sh -d "$ARG_BUILD_DIR" -r ${gitrepo} -x
   else
-    ./generate_documents.sh -d $ARG_BUILD_DIR -r ${gitrepo} -b $branch
+    ./generate_documents.sh -d "$ARG_BUILD_DIR" -r "${gitrepo}" -b "$branch" -x
   fi
 done
+# one last time without supress flag
+# this last run builds the statics via "npm run build"
+./generate_documents.sh -d "$ARG_BUILD_DIR" -r "eosnetworkfoundation/mandel-swift"
